+++ python/sglang/jit_kernel/tests/test_fused_store_index_cache.py
@@ -187,7 +187,7 @@ def _reference_quantize_and_store(
 
 def _import_act_quant():
     try:
-        from sglang.srt.layers.attention.nsa.triton_kernel import act_quant
+        from sglang.srt.layers.attention.dsa.triton_kernel import act_quant
 
         return act_quant
     except Exception:
