+++ python/sglang/srt/layers/attention/base_attn_backend.py
@@ -9,7 +9,7 @@ from sglang.kernel_api_logging import debug_kernel_api
 from sglang.srt.utils.common import is_npu
 
 if TYPE_CHECKING:
-    from sglang.srt.layers.attention.nsa.nsa_indexer import BaseIndexerMetadata
+    from sglang.srt.layers.attention.dsa.nsa_indexer import BaseIndexerMetadata
     from sglang.srt.layers.radix_attention import RadixAttention
     from sglang.srt.model_executor.forward_batch_info import ForwardBatch, ForwardMode
     from sglang.srt.speculative.spec_info import SpecInput
