+++ python/sglang/srt/layers/communicator.py
@@ -33,7 +33,7 @@ from sglang.srt.distributed.device_communicators.pynccl_allocator import (
     use_symmetric_memory,
 )
 from sglang.srt.environ import envs
-from sglang.srt.layers.attention.nsa.utils import (
+from sglang.srt.layers.attention.dsa.utils import (
     is_nsa_enable_prefill_cp,
     nsa_use_prefill_cp,
 )
